Update tpg_tcp_sm.c
There is a bug where in the timers are not cancelled, when the acknowledgement arrives for the data segment, which causes the timer queues to keep growing (and also re-transmissions). The proposed fix solves the issue. 

Without this fix, following counters can verify the existing issue. 
Setup : establish 1 connection with a test duration of say 2 seconds, in a back to back setup with port 0 as part of client profile and port 1 as part of server profile.  
There after issue 
warp17> show timer statistics
Port 0 TCP Timer statistics:
  RTO Timer Set       :                  356
  RTO Timer Cancelled :                   1
  RTO Timer Fired     :                  354

Port 1 TCP Timer statistics:
  RTO Timer Set       :                  356
  RTO Timer Cancelled :                   1
  RTO Timer Fired     :                  354